refactor(models): change price field from DecimalField to IntegerField
The price field in the Service model has been changed from DecimalField
to IntegerField for simplification. This affects how prices are stored
and displayed, removing decimal precision in favor of integer values.

BREAKING CHANGE: Price field now stores integer values instead of
decimal values, which may affect existing price data and calculations.

+++ app_home/models.py
@@ -4,7 +4,7 @@ from django.contrib.auth.models import User
 class Service(models.Model):
     name = models.CharField(max_length=200)
     description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2, verbose_name="Цена")
+    price = models.IntegerField(verbose_name="Цена")
 
     class Meta:
         verbose_name = "Услуга"
